Add files via upload
This repository contains PowerShell scripts for detecting and remediating the deprecation vulnerability associated with TLS 1.0 and TLS 1.1. These legacy protocols are considered insecure and must be explicitly disabled to comply with modern security standards and best practices, such as PCI DSS and NIST recommendations.

The scripts ensure that both the Client and Server configurations for TLS 1.0 and TLS 1.1 are explicitly disabled by modifying the relevant Windows registry keys.